firebase_migration/functions/requirements.txt
firebase-admin==6.4.0
firebase-functions==0.4.0
pinecone-client==3.0.0
openai==1.10.0
